change some logic for concurrency
